+++ lib/src/configuration.dart
@@ -645,8 +645,8 @@ class SyncError extends RealmError {
         return SyncConnectionError(message, category, SyncConnectionErrorCode.fromInt(code), isFatal: isFatal);
       case SyncErrorCategory.session:
         return SyncSessionError(message, category, SyncSessionErrorCode.fromInt(code), isFatal: isFatal);
-      case SyncErrorCategory.resolve:
-        return SyncResolveError(message, category, SyncResolveErrorCode.fromInt(code));
+      case SyncErrorCategory.webSocket:
+        return SyncWebSocketError(message, category, SyncWebSocketErrorCode.fromInt(code));
       case SyncErrorCategory.system:
       case SyncErrorCategory.unknown:
       default:
@@ -729,16 +729,16 @@ class SyncSessionError extends SyncError {
   }
 }
 
-/// Network resolution error
-class SyncResolveError extends SyncError {
-  /// The numeric value indicating the type of the network resolution sync error.
-  SyncResolveErrorCode get code => SyncResolveErrorCode.fromInt(codeValue);
+/// Web socket error
+class SyncWebSocketError extends SyncError {
+  /// The numeric value indicating the type of the web socket sync error.
+  SyncWebSocketErrorCode get code => SyncWebSocketErrorCode.fromInt(codeValue);
 
-  SyncResolveError(String message, SyncErrorCategory category, SyncResolveErrorCode errorCode) : super(message, category, errorCode.index);
+  SyncWebSocketError(String message, SyncErrorCategory category, SyncWebSocketErrorCode errorCode) : super(message, category, errorCode.index);
 
   @override
   String toString() {
-    return "SyncResolveError message: $message category: $category code: $code";
+    return "SyncWebSocketError message: $message category: $category code: $code";
   }
 }